basic user tests added & user entity fix
- added basic tests for user dao
- user entity fixed so that tests can work
- type of id in monster changed from Long to int

+++ src/main/java/cz/muni/fi/pa165/dao/MonsterDao.java
@@ -31,7 +31,7 @@ public interface MonsterDao {
 	 * @param id id of a monster
 	 * @return found monster
 	 */
-	Monster findById(Long id);
+	Monster findById(int id);
 
 	/**
 	 * Finds monster by its type.
